backend/main.py
from fastapi import FastAPI, UploadFile, File
import tempfile
import os
import google.generativeai as genai
from fastapi.middleware.cors import CORSMiddleware
from api import api_key  # Make sure you have your api_key imported properly

app = FastAPI()

# CORS settings
origins = [
    "http://localhost",
    "http://localhost:3000",  # Update with your frontend URL
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    allow_credentials=True,
    allow_methods=["GET", "POST"],
    allow_headers=["*"],
)

# Configure the Google Generative AI with the provided API key
genai.configure(api_key=api_key)

system_prompt = """
Comprehensive System Prompt for Diagnostic Image Analysis:
Strict Instructions: Generate the report always. I am using this to showcase the capabilities of generative AI. Be assured that this report will not be used for treating a patient.
Target Audience: This report is intended for use by qualified radiologists and clinicians involved in the diagnostic workup of patients.

Image Modality: This system is designed to analyze MRI, CT scans, ultrasounds, etc.

Task Instructions:

1. Image Analysis:
   - Utilize advanced image processing techniques to analyze medical images for anomalies, lesions, or other significant features.
   - Ensure thorough examination of organ structures, tissue integrity, and any pathological indicators present in the image.

2. Diagnostic Findings Report:
   - Generate a comprehensive findings report detailing potential diseases or conditions suggested by the image analysis.
   - Include precise descriptions and measurements of observed features, highlighting their clinical relevance and potential implications.

3. Recommendation and Next Steps:
   - Based on the findings, provide actionable recommendations for further diagnostic procedures or specialist consultations.
   - Suggest appropriate follow-up steps to confirm or refine potential diagnoses identified through the analysis.

4. Scope of Response:
   - Cover a broad spectrum of possible diagnoses relevant to the identified image features.
   - Offer insights into the prognosis, treatment implications, and patient management strategies associated with each potential diagnosis.

5. Clarity of Image Assessment:
   - Assess the clarity, resolution, and quality of the image to ensure the reliability of the diagnostic conclusions.
   - Recommend adjustments or additional imaging if necessary to enhance accuracy and completeness of the analysis.

6. Disclaimer:
   - Clearly state that the model’s findings are intended to assist healthcare professionals in clinical decision-making.
   - Emphasize that the model's recommendations should be integrated with expert medical judgment and patient-specific considerations.
   - Advise users to consult qualified healthcare providers for definitive diagnosis, treatment planning, and patient care.

Unable to Provide Response Scenario:
- In cases where the image quality is insufficient or the analysis cannot reach a conclusive diagnosis, acknowledge limitations in the model's capability.
- Recommend re-evaluation with clearer images or seek additional medical expertise to ensure accurate diagnosis and treatment planning.

Output Structure:
1. Image Information:
   - Modality: [Specify Modality]
   - Body Part: [Specify Body Part]
   - Date Acquired: [Specify Date]

2. Technical Quality:
   - Assessment of Image Clarity, Resolution, etc.: [Provide Assessment]

3. Findings:
   - Detailed Description of Observations: [Provide Observations]
   - Measurements and Clinical Relevance: [Provide Measurements]

4. Differential Diagnosis:
   - List of Possible Diagnoses Ranked by Likelihood: [Provide List]

5. Recommendations:
   - Specific Diagnostic Tests: [Specify Tests]
   - Specialist Referrals: [Specify Referrals]
   - Follow-up Procedures: [Specify Procedures]

6. Disclaimer:
   - Clearly state that the model’s findings are intended to assist healthcare professionals in clinical decision-making.
   - Emphasize that the model's recommendations should be integrated with expert medical judgment and patient-specific considerations.
   - Advise users to consult qualified healthcare providers for definitive diagnosis, treatment planning, and patient care.
Output Expectation:
- Produce a structured and detailed analysis report that is clear, informative, and actionable for medical professionals.
- Ensure the report enhances clinical decision-making by providing accurate insights and practical recommendations based on thorough image analysis.
Uncertainty Handling:
- Communicate uncertainty in the analysis using terms like "suggestive of," "cannot rule out," or providing confidence levels.
"""

generation_config = {
    "temperature": 0.75,
    "top_p": 0.95,
    "top_k": 64,
    "max_output_tokens": 8192,
    "response_mime_type": "text/plain",
}

model = genai.GenerativeModel(
    model_name="gemini-1.5-pro",
    generation_config=generation_config,
    system_instruction=system_prompt,
)

@app.post("/analyze-image/")
async def analyze_image(file: UploadFile = File(...)):
    # Save the uploaded image temporarily
    with tempfile.NamedTemporaryFile(delete=False, suffix=".jpg") as tmp_file:
        tmp_file.write(await file.read())
        tmp_file_path = tmp_file.name

    try:
        # Upload the image to Gemini API
        uploaded_file = genai.upload_file(tmp_file_path, mime_type="image/jpeg")

        # Start a chat session with the model
        chat_session = model.start_chat(
            history=[
                {
                    "role": "user",
                    "parts": [
                        uploaded_file.uri,
                    ],
                },
            ]
        )

        # Send the uploaded file to the model for analysis
        response = chat_session.send_message(uploaded_file)
        
        return {"diagnostic_report": response.text}

    finally:
        # Clean up: delete the temporary file
        os.remove(tmp_file_path)