backend/requirements.txt
fastapi==0.74.*
uvicorn[standard]==0.17.*
google-generativeai
python-multipart